+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `submissionTitle` to `approvalsService.getFormSubmissionApprovals()`
+- `submissionTitle` to `approvalsService.getFormSubmissionAdministrationApprovals()`
+
 ## [6.1.0] - 2023-08-01
 
 ### Added